requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.4